move answer set to Question class
